Remove circular dependency between mugshot and webdriverio
The mugshot package depends on the @mughsot/webdriverio package for the
gui tests, but lerna complains about the circular dependency. In
practice it's not needed because of how module hoisting works.

This sparks the question of whether mugshot should really use one
particular implementation of the `Browser` interface in its tests.
Should it always use a mock, thus making the gui tests not e2e?
Should those tests not exist at all because they're overlapping
with the node tests anyway?